fix escope patch and improve tests
