+++ lib/analyze-scope.js
@@ -322,7 +322,7 @@ module.exports = function(ast, parserOptions) {
   };
 
   if (OriginalReferencer._babelEslintPatched) {
-    options._skipBabelEslintGlobals = true;
+    require("./patch-eslint-scope")(parserOptions);
     return escope.analyze(ast, options);
   }
