+++ package.json
@@ -20,7 +20,7 @@
   },
   "scripts": {
     "test": "npm run lint && npm run test-only",
-    "test-only": "mocha",
+    "test-only": "mocha && mocha --require test/fixtures/preprocess-to-patch.js && mocha --require test/fixtures/use-eslint-old.js",
     "lint": "eslint index.js babylon-to-espree test",
     "fix": "eslint index.js babylon-to-espree test --fix",
     "precommit": "lint-staged",
