+++ test/fixtures/preprocess-to-patch.js
@@ -1,0 +1,5 @@
+"use strict"
+const babelEslint = require("../..")
+
+// Apply monkeypatch to eslint-scope.
+babelEslint.parse("var x = 0;")
